Ajout de ServerUnitSpawnPacket Mise a jour de l'enum de APacket (réorganisation et ajout de 4 valeurs pour server) Mise a jour de tout le code pour que ça compile sur Mac. On met maintenant  #if (defined __linux__) || (defined __APPLE__) à la place de juste linux svp Mise a jour de Timer (passage de uintmax_t à uint64_t)

+++ srcs/misc/network/Networker.cpp
@@ -83,7 +83,7 @@ Networker::~Networker()
 void Networker::send(APacket *pack, int id)
 {
   std::string data = pack->getData();
-  int dataSize = data.size();
+  unsigned long dataSize = data.size();
 
   _sock->setSendPort(_peers[id].second);
   _sock->setAddr(_peers[id].first);
